LayoutTests/fast/css/font-face-descriptor-multiple-values-parsing-expected.txt

This tests that the font-weight, font-style and font-variant properties accept value lists when they appear inside @font-face rules, and that the value 'all' is allowed only by itself.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("font-weight", "100") is "100"
PASS test("font-weight", "all") is "all"
PASS test("font-weight", "100, 200") is "100, 200"
PASS test("font-weight", "bold, normal") is "bold, normal"
PASS test("font-weight", "100, 200, 300, 400, 500, 600, 700, 100") is "100, 200, 300, 400, 500, 600, 700, 100"
PASS test("font-weight", "all, 100") threw exception TypeError: Null value.
PASS test("font-weight", "bold, normal, all") threw exception TypeError: Null value.
PASS test("font-weight", "") threw exception TypeError: Null value.
PASS test("font-style", "normal") is "normal"
PASS test("font-style", "italic") is "italic"
PASS test("font-style", "normal, oblique") is "normal, oblique"
PASS test("font-style", "all") is "all"
PASS test("font-style", "all, normal") threw exception TypeError: Null value.
PASS test("font-style", "italic, all") threw exception TypeError: Null value.
PASS test("font-style", "") threw exception TypeError: Null value.
PASS test("font-variant", "normal") is "normal"
PASS test("font-variant", "small-caps") is "small-caps"
PASS test("font-variant", "normal, small-caps") is "normal, small-caps"
PASS test("font-variant", "all") is "all"
PASS test("font-variant", "all, normal") threw exception TypeError: Null value.
PASS test("font-variant", "small-caps, all") threw exception TypeError: Null value.
PASS test("font-variant", "") threw exception TypeError: Null value.
PASS successfullyParsed is true

TEST COMPLETE
